Test for UI (Event/Cycle and Meters) + Small changes on main scene (Button Pause & more)
Change of the color of many panel + New pause Button + moving FPS and Moods not to be on the center of the screen on the main scene.
Modification of the Piano Building by adding the Script for impact on meters.
Modification of the Value for T building no real issue apparently.
Test in my folder to prepare Event/Cycle and Meters UI. (Adding many sprites)